feat(LogViewer): refactor toolbar functionality
