+++ packages/react-log-viewer/src/LogViewer/LogViewerContext.tsx
@@ -7,3 +7,17 @@ export interface LogViewerContextInterface {
 }
 
 export const LogViewerContext = createContext<LogViewerContextInterface | null>(null);
+
+interface LogViewerToolbarContextProps {
+  searchedWordIndexes: number[];
+  rowInFocus: number;
+  searchedInput: string;
+  currentSearchedItemCount: number;
+  scrollToRow: (searchedRow: number) => void;
+  setRowInFocus: (index: number) => void;
+  setSearchedInput: (input: string) => void;
+  setSearchedWordIndexes: (indexes: number[]) => void;
+  setCurrentSearchedItemCount: (index: number) => void;
+}
+
+export const LogViewerToolbarContext = createContext<LogViewerToolbarContextProps | null>(null);
