+++ packages/react-log-viewer/src/LogViewer/examples/LogViewer.md
@@ -2,50 +2,62 @@
 id: Log viewer
 beta: true
 section: components
-propComponents: [LogViewer, LogViewerCustomControl]
+propComponents: [LogViewer, LogViewerSearch]
 ---
 
-import { LogViewer, LogViewerCustomControl } from '@patternfly/react-log-viewer';
-
+import { LogViewer, LogViewerSearch } from '@patternfly/react-log-viewer';
 import { bigData } from './bigTestData.js';
-
 import { data } from './realTestData.js';
 
 ## Examples
-### Basic log-viewer
+
+### Basic
+
 ```js
 import React from 'react';
 import { data } from './data/realTestData.js';
 import { LogViewer } from '@patternfly/react-log-viewer';
 
-<LogViewer hasLineNumbers={false} hasToolbar={false} height={300}/>
+<LogViewer hasLineNumbers={false} height={300} />;
 ```
 
-### Log-viewer with toolbar and data
+### With search
+
 ```js
 import React from 'react';
 import { data } from './realTestData';
 import { bigData } from './bigTestData.js';
-import { LogViewer } from '@patternfly/react-log-viewer';
+import { LogViewer, LogViewerSearch } from '@patternfly/react-log-viewer';
+import { Toolbar, ToolbarContent, ToolbarItem, ToolbarGroup } from '@patternfly/react-core';
 
 BasicLogViewer = () => {
-  return(
-    <React.Fragment>
-      <LogViewer data={data.data} />
-    </React.Fragment>
-  );  
+  return (
+    <LogViewer
+      data={data.data}
+      toolbar={
+        <Toolbar>
+          <ToolbarContent>
+            <ToolbarItem>
+              <ToolbarGroup>
+                <LogViewerSearch placeholder="Search value" />
+              </ToolbarGroup>
+            </ToolbarItem>
+          </ToolbarContent>
+        </Toolbar>
+      }
+    />
+  );
 };
 ```
 
-### Log-viewer with custom button
+### With complex toolbar
+
 ```js
 import React from 'react';
 import { data } from './realTestData';
 import { bigData } from './bigTestData.js';
-import { LogViewer } from '@patternfly/react-log-viewer';
-import { LogViewerCustomControl } from '@patternfly/react-log-viewer';
-import { Dropdown, DropdownItem, DropdownToggle } from '@patternfly/react-core';
-import PlayIcon from '@patternfly/react-icons/dist/js/icons/play-icon';
+import { LogViewer, LogViewerSearch } from '@patternfly/react-log-viewer';
+import { Toolbar, ToolbarContent, ToolbarItem, ToolbarGroup, Button } from '@patternfly/react-core';
 
 CustomControlLogViewer = () => {
   const [isOpen, setIsOpen] = React.useState(false);
@@ -53,19 +65,39 @@ CustomControlLogViewer = () => {
   const dataSourcesAmount = 2;
   const currentDataSource = React.useState(0);
 
-  const onClickExample = (event, value) => {
-    console.log('MY VALUE', value); //eslint-disable-line
+  const onActionClick = event => {
+    console.log('clicked test action button');
   };
 
-  const customControls = (
-    <LogViewerCustomControl onClick={onClickExample}>Print to Console</LogViewerCustomControl>
+  const onPrintClick = event => {
+    console.log('clicked console print button');
+  };
+
+  return (
+    <LogViewer
+      data={data.data}
+      toolbar={
+        <Toolbar>
+          <ToolbarContent>
+            <ToolbarItem>
+              <Button onClick={onActionClick} variant="control">
+                Test Action
+              </Button>
+            </ToolbarItem>
+            <ToolbarItem>
+              <ToolbarGroup>
+                <LogViewerSearch placeholder="Search" />
+              </ToolbarGroup>
+            </ToolbarItem>
+            <ToolbarItem>
+              <Button onClick={onPrintClick} variant="control">
+                Print to Console
+              </Button>
+            </ToolbarItem>
+          </ToolbarContent>
+        </Toolbar>
+      }
+    />
   );
-
-  return(
-    <React.Fragment>
-      <LogViewer data={data.data} customControls={customControls} />
-    </React.Fragment>
-  );  
 };
-
 ```
